Update README.md
Beginning in Windows PowerShell 5.0, you can install multiple versions of the same module on the same computer; even in the same directory. Open source and PowerShellGet have revolutionized the availability of modules and Windows PowerShell 5.0+ will be continuously updated with Windows. 

The result is a myriad of interlocking parts with far more potential for conflicts in name, version, and functionality. 

When you run a command, how do you guarantee that you're running the command that you intend, instead of a command with the same name in a different module or a different version of the same module? 

How do you manage conflicting requirements for versions of PowerShell and versions of modules? 

Are we fated for the old "DLL Hell?" 

In this talk, I'll present the problem, describe some mitigating strategies, warn about their limitations, and provide a roadmap for version sanity.